Moved Preferences to the top level for use by non-GUI code (CLIs, etc).
